Add new Taskmaster

+++ src/process.js
@@ -1,38 +1,42 @@
 
 const spawn               = require('child_process').spawn;
 const fs                  = require('fs');
+const json 				  = require('../config/config.json');
 
 module.exports            = class Process {
-  constructor(process_config, process_name, taskmaster, old_path) {
+  constructor(process_config, process_name, taskmaster, old_path, retries) {
 
     this.taskmaster       = taskmaster;
     this.name             = process_name;
     this.args             = process_config.cmd.split(' ');
     this.cmd_bis          = this.args[0];
     this.old_path         = old_path;
-    this.state            = 'stopped';
+    this.state            = 'stopped'
     this.uptime           = 0;
-    this.retries          = 0;
-   
+    if (retries)
+    	this.retries      = retries;
+    else
+    	this.retries      = 0;
     this.args.splice(0, 1);
     this.reload_config(process_config);
   }
 
   reload_config(process_config) {
     this.cmd              = process_config.cmd;
-    this.numprocs         = process_config.numprocs || 1;
-    this.umask            = process_config.umask || '022';
-    this.workingdir       = process_config.workingdir || '/tmp';
-    this.autostart        = process_config.autostart || true;
-    this.autorestart      = process_config.autorestart || 'unexpected';
-    this.exitcodes        = process_config.exitcodes || [0];
-    this.startretries     = process_config.startretries || 3;
-    this.starttime        = process_config.starttime || 5;
-    this.stopsignal       = process_config.stopsignal || 'Kill';
-    this.stoptime         = process_config.stoptime || 10;
+    this.numprocs         = process_config.numprocs;
+    this.umask            = process_config.umask;
+    this.workingdir       = process_config.workingdir;
+    this.autostart        = process_config.autostart;
+    this.autorestart      = process_config.autorestart;
+    this.exitcodes        = process_config.exitcodes;
+    this.startretries     = process_config.startretries;
+    this.starttime        = process_config.starttime;
+    this.stopsignal       = process_config.stopsignal;
+    this.stoptime         = process_config.stoptime;
     this.stdout           = process_config.stdout;
     this.stderr           = process_config.stderr;
-    this.env              = process_config.env || {};
+    this.env              = process_config.env;
+    this.type 			  = process_config.type;
     this.spawn_process();
   }
 
@@ -45,7 +49,6 @@ module.exports            = class Process {
     let args              = this.cmd.split(' ');
     let process_cmd       = args[0];
     args.splice(0, 1);
-
 
     if (!fs.existsSync(process_cmd))
       this.taskmaster.logger.error(`Cannot launch process ${this.name}`);
@@ -64,6 +67,7 @@ module.exports            = class Process {
 
     this._process.stdout.on('data', this._on_stdout.bind(this));
     this._process.stderr.on('data', this._on_stderr.bind(this));
+
     this.pid               = this._process.pid;
     setTimeout(() => {
       if (this.is_running()) {
@@ -71,9 +75,10 @@ module.exports            = class Process {
         this._process.on('close', this._on_close.bind(this));
         this.taskmaster.logger.info(`\x1b[32m${this.name} has sucessfully started.\x1b[0m`);
       } 
-      else
+      else {
         this._on_close(this._process.exitCode);
-    }, 1000 * this.startime);
+    	}
+    }, 1000 * this.starttime);
   }
 
   set_std() {
@@ -112,7 +117,7 @@ module.exports            = class Process {
   }
 
   _on_close(code) {
-    this.pid                = null;
+    // this.pid                = null;
     this.state              = 'stopped';
     if (this.stdout_stream)
       this.stdout_stream.end();
@@ -124,16 +129,20 @@ module.exports            = class Process {
       this.taskmaster.logger.info(`\x1b[32mProcess ${this.name} has normally stopped with code ${code}.\x1b[0m`);
       return;
     }
-    this.taskmaster.logger.warn(`\x1b[31mProcess ${this.name} has crashed with code ${code}\x1b[0m`);
-    if (this.retries < this.startretries && !this.startretries != 'never') {
-      this.restart();
-      this.retries++;
+    this.taskmaster.logger.warn(`\x1b[31mProcess ${this.name} has crashed with code unexpected.\x1b[0m`);    console.log(`${this.retries} < ${this.startretries}`)
+    if (this.retries < this.startretries && this.startretries != 'never') {
+  		this.retries++;
+  		this.state = 'started';
+      	this.restart();
     }
+    else 
+    	this.pid  = null;
   }
 
   is_running() {
     try {
-      return process.kill(this.pid,0);
+    	if (this.type == "loop")
+      		return process.kill(this.pid,0);
     }
     catch(e) {
       this.taskmaster.logger.error(`\x1b[31mError : ${e}\x1b[0m`);
@@ -141,7 +150,7 @@ module.exports            = class Process {
   }
 
   stop(force_stop = false) {
-   if (force_stop == true)
+   if (force_stop == true)''
       this.startretries = 0;
     this.uptime = 0;
     try {
@@ -155,16 +164,17 @@ module.exports            = class Process {
 
   restart() {
     this.stop()
-    new Process (
+    let restart = new Process (
       this.taskmaster.config.options[this.name],
       this.name,
       this.taskmaster,
-      this.old_path);
-    this.taskmaster.logger.info(`\n\x1b[32m${this.name} : started\x1b[0m`);
+      this.old_path,
+      this.retries);
+    this.taskmaster.process_manager.processes[this.name].push(restart);
+    this.taskmaster.logger.info(`\x1b[32mRestaring : ${this.name} + ${this.pid} + ${this._process.pid}\x1b[0m`);
   }
 
   status(process_name, index) {
-
     if (this._process.pid != null && this.state == 'started')
       console.log(`[${process_name}][${index}] RUNNING with [${this._process.pid}] and uptime ${this.uptime}s`);
     else
